python/requirements.txt
tqdm >= 4.0.0
mutagen >= 1.46.0
opencv-python >= 4.12.0.88
pillow >= 12.0.0